+++ mm/CMakeLists.txt
@@ -140,6 +140,7 @@ file(GLOB_RECURSE soh__DeveloperTools RELATIVE ${CMAKE_CURRENT_SOURCE_DIR}
    "2s2h/DeveloperTools/*.h"
    "2s2h/DeveloperTools/*.c"
    "2s2h/DeveloperTools/*.cpp"
+   "2s2h/DeveloperTools/*.hpp"
 )
 # }}}
 
@@ -158,6 +159,13 @@ file(GLOB_RECURSE soh__Enhancements RELATIVE ${CMAKE_CURRENT_SOURCE_DIR}
     "2s2h/Enhancements/*.hpp"
     "2s2h/Enhancements/*_extern.inc"
     "2s2h/Enhancements/*.mm"
+)
+
+file(GLOB_RECURSE soh__BenGui RELATIVE ${CMAKE_CURRENT_SOURCE_DIR}
+    "2s2h/BenGui/*.c"
+    "2s2h/BenGui/*.cpp"
+    "2s2h/BenGui/*.h"
+    "2s2h/BenGui/*.hpp"
 )
 
 #list(REMOVE_ITEM soh__Enhancements "soh/Enhancements/gamecommand.h")
@@ -181,8 +189,7 @@ file(GLOB_RECURSE soh__Enhancements RELATIVE ${CMAKE_CURRENT_SOURCE_DIR}
 
 #source_group("soh\\Enhancements" REGULAR_EXPRESSION "soh/Enhancements/*")
 #source_group("soh\\Enhancements\\audio" REGULAR_EXPRESSION "soh/Enhancements/audio/*")
-source_group("2s2h\\Enhancements\\controls" REGULAR_EXPRESSION "2s2h/Enhancements/controls/*")
-source_group("2s2h\\Enhancements\\interpolation" REGULAR_EXPRESSION "2s2h/Enhancements/interpolation/*")
+# source_group("2s2h\\Enhancements\\interpolation" REGULAR_EXPRESSION "2s2h/Enhancements/interpolation/*")
 #source_group("soh\\Enhancements\\cosmetics" REGULAR_EXPRESSION "soh/Enhancements/cosmetics/*")
 #source_group("soh\\Enhancements\\crowd-control" REGULAR_EXPRESSION "soh/Enhancements/crowd-control/*")
 #source_group("soh\\Enhancements\\custom-message" REGULAR_EXPRESSION "soh/Enhancements/custom-message/*")
@@ -272,6 +279,7 @@ set(ALL_FILES
     ${soh__DeveloperTools}
 #    ${soh__config}
     ${soh__Enhancements}
+    ${soh__BenGui}
     ${soh__Extractor}
     ${soh__Resource}
     ${src__}
