+++ mm/src/code/sys_matrix.c
@@ -41,7 +41,7 @@
 
 #include "prevent_bss_reordering.h"
 #include "global.h"
-#include "2s2h/Enhancements/interpolation/frame_interpolation.h"
+#include "2s2h/Enhancements/FrameInterpolation/FrameInterpolation.h"
 
 /* data */
 #define qs1616(e) ((s32)((e)*0x00010000))
