+++ mm/src/code/z_eff_spark.c
@@ -1,6 +1,6 @@
 #include "global.h"
 #include "objects/gameplay_keep/gameplay_keep.h"
-#include "2s2h/Enhancements/interpolation/frame_interpolation.h"
+#include "2s2h/Enhancements/FrameInterpolation/FrameInterpolation.h"
 
 void EffectSpark_Init(void* thisx, void* initParamsx) {
     EffectSpark* this = (EffectSpark*)thisx;
