+++ mm/src/code/z_lights.c
@@ -2,7 +2,7 @@
 #include "sys_cfb.h"
 #include "objects/gameplay_keep/gameplay_keep.h"
 #include <string.h>
-#include "2s2h/Enhancements/interpolation/frame_interpolation.h"
+#include "2s2h/Enhancements/FrameInterpolation/FrameInterpolation.h"
 
 LightsBuffer sLightsBuffer;
