+++ mm/src/code/z_vr_box_draw.c
@@ -1,5 +1,5 @@
 #include "global.h"
-#include "2s2h/Enhancements/interpolation/frame_interpolation.h"
+#include "2s2h/Enhancements/FrameInterpolation/FrameInterpolation.h"
 
 Mtx* sSkyboxDrawMatrix;
