+++ mm/src/overlays/actors/ovl_Object_Kankyo/z_object_kankyo.c
@@ -6,7 +6,7 @@
 
 #include "z_object_kankyo.h"
 #include "objects/gameplay_keep/gameplay_keep.h"
-#include "2s2h/Enhancements/interpolation/frame_interpolation.h"
+#include "2s2h/Enhancements/FrameInterpolation/FrameInterpolation.h"
 
 #define FLAGS (ACTOR_FLAG_10 | ACTOR_FLAG_20 | ACTOR_FLAG_2000000)
